Moved all ServerCommands to SlashCommands

+++ cogs/servercommands.py
@@ -1,9 +1,10 @@
 import discord
+from discord import Option, slash_command
+
 import utils.globals as GG
 
 from discord.ext import commands
 from crawler_utilities.handlers import logger
-from crawler_utilities.utils.functions import try_delete
 from utils import checks
 
 log = logger.logger
@@ -12,7 +13,7 @@ categories = ['ANON', 'DELIVERY', 'MODDECISION', 'BLACKLIST']
 
 
 def getRole(roleID, ctx):
-    guild = ctx.message.guild
+    guild = ctx.guild
     role = guild.get_role(roleID)
     return role.name
 
@@ -31,54 +32,55 @@ class ServerCommands(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.command()
-    @GG.is_staff()
-    @commands.guild_only()
+    @slash_command(name="addstaff")
     async def addstaff(self, ctx, role: discord.Role):
-        """[STAFF ONLY]"""
+        """[STAFF] Adds a role to the staff list"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
+
         if role.id in GG.STAFF:
-            await ctx.send(f"{role.name} is already on the staff list.")
+            return await ctx.respond(f"{role.name} is already on the staff list.", ephemeral=True)
         else:
             await GG.MDB['serverstaff'].insert_one({"guild": ctx.guild.id, "roles": int(role.id)})
             GG.STAFF.append(int(role.id))
-            await ctx.send(f"{role.name} was added to the staff list.")
+            return await ctx.respond(f"{role.name} was added to the staff list.", ephemeral=True)
 
-    @commands.command()
-    @GG.is_staff()
-    @commands.guild_only()
+    @slash_command(name="stafflist")
     async def stafflist(self, ctx):
-        """[STAFF ONLY]"""
-        user_quotes = await GG.MDB['serverstaff'].find({"guild": ctx.message.guild.id}).to_list(length=None)
-        if user_quotes is not None:
-            await ctx.send(embed=list_embed(user_quotes, ctx.author, ctx))
-        else:
-            await ctx.send("This server has no roles assigned as staff.")
+        """[STAFF] Returns a list of all staff roles"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
 
-    @commands.command()
-    @GG.is_staff()
-    @commands.guild_only()
+        user_quotes = await GG.MDB['serverstaff'].find({"guild": ctx.guild.id}).to_list(length=None)
+        if user_quotes is not None:
+            return await ctx.respond(embed=list_embed(user_quotes, ctx.author, ctx), ephemeral=True)
+        else:
+            return await ctx.respond("This server has no roles assigned as staff.", ephemeral=True)
+
+    @slash_command(name="delstaff")
     async def delstaff(self, ctx, role: discord.Role):
-        """[STAFF ONLY]"""
+        """[STAFF] Removes a role to the staff list"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
+
         exist = await GG.MDB['serverstaff'].find_one({"guild": ctx.guild.id, "roles": int(role.id)})
         if exist is not None:
             await GG.MDB['serverstaff'].delete_one({"guild": ctx.guild.id, "roles": int(role.id)})
             GG.STAFF.remove(int(role.id))
-            await ctx.send(f"{role.name} was removed from the staff list.")
+            return await ctx.respond(f"{role.name} was removed from the staff list.", ephemeral=True)
         else:
-            await ctx.send(f"{role.name} is not on the staff list.")
+            return await ctx.respond(f"{role.name} is not on the staff list.", ephemeral=True)
 
-    @commands.command()
-    @commands.guild_only()
-    @checks.admin_or_permissions(manage_guild=True)
-    async def prefix(self, ctx, prefix: str = None):
-        """Sets the bot's prefix for this server.
-        Forgot the prefix? Reset it with "@DiscordCrawler#6716 prefix $".
-        """
-        await try_delete(ctx.message)
+    @slash_command(name="prefix")
+    async def prefix(self, ctx, prefix: Option(str, "What do you want the prefix to be")):
+        """[SERVER OWNER] Sets the bot's prefix for this server. Forgot the prefix? Reset it with "/prefix $"""
+        if not checks.admin_or_permissions(manage_guild=True):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
+
         guild_id = str(ctx.guild.id)
         if prefix is None:
             current_prefix = await self.bot.get_server_prefix(ctx.message)
-            return await ctx.send(f"My current prefix is: `{current_prefix}`")
+            return await ctx.respond(f"My current prefix is: `{current_prefix}`", ephemeral=True)
 
         self.bot.prefixes[guild_id] = prefix
 
@@ -88,35 +90,28 @@ class ServerCommands(commands.Cog):
             upsert=True
         )
 
-        await ctx.send("Prefix set to `{}` for this server.".format(prefix))
+        return await ctx.respond("Prefix set to `{}` for this server.".format(prefix), ephemeral=True)
 
-    @commands.command()
-    @GG.is_staff()
-    @commands.guild_only()
-    async def addchannel(self, ctx, TYPE: str, channel: discord.TextChannel):
-        """[STAFF ONLY]"""
-        if GG.checkPermission(ctx, "mm"):
-            TYPE = TYPE.upper()
-            if TYPE in categories:
-                channelExists = await GG.MDB['channelinfo'].find_one({"channel": channel.id})
-                if channelExists is not None:
-                    await ctx.send(f"{channel.name} is already on the special channel list.")
-                else:
-                    await GG.MDB['channelinfo'].insert_one({"guild": ctx.guild.id, "channel": channel.id, "type": TYPE})
-                    GG.CHANNEL[channel.id] = str(TYPE)
-                    await ctx.send(f"{channel.name} was added to the special channel list as {TYPE}")
-            else:
-                await ctx.send(
-                    f"{TYPE} is not part of the list of specialized channels you can make. Please use one of the following:\n{categories}")
+    @slash_command(name="addchannel")
+    async def addchannel(self, ctx, channeltype: Option(str, "Choose your channel type", choices=categories), channel: Option(discord.TextChannel, "For which channel do you want to set the type?")):
+        """[STAFF] Gives a channel a specific type."""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
+
+        channelExists = await GG.MDB['channelinfo'].find_one({"channel": channel.id})
+        if channelExists is not None:
+            return await ctx.respond(f"{channel.name} is already on the special channel list.", ephemeral=True)
         else:
-            await ctx.send(
-                "I don't have the Manage_Messages permission. It's a mandatory permission, I have noted my owner about this. Please give me this permission, I will end up leaving the server if it happens again.")
+            await GG.MDB['channelinfo'].insert_one({"guild": ctx.guild.id, "channel": channel.id, "type": channeltype})
+            GG.CHANNEL[channel.id] = str(channeltype)
+            return await ctx.respond(f"{channel.name} was added to the special channel list as {channeltype}", ephemeral=True)
 
-    @commands.command()
-    @GG.is_staff()
-    @commands.guild_only()
-    async def delchannel(self, ctx, channel: discord.TextChannel):
-        """[STAFF ONLY]"""
+    @slash_command(name="delchannel")
+    async def delchannel(self, ctx, channel: Option(discord.TextChannel, "From what channel do you want to remove the channel type?")):
+        """[STAFF] Removes the type of a channel, reverting it to a "normal" channel"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
+
         channelExists = await GG.MDB['channelinfo'].find_one({"channel": channel.id})
         if channelExists is not None:
             await GG.MDB['channelinfo'].delete_one({"channel": channel.id})
@@ -124,14 +119,6 @@ class ServerCommands(commands.Cog):
             await ctx.send(f"{channel.name} was removed from the special channel list.")
         else:
             await ctx.send(f"{channel.name} is not on the special channel list.")
-
-    @commands.command()
-    @GG.is_staff()
-    @commands.guild_only()
-    async def channeltypes(self, ctx):
-        """[STAFF ONLY]"""
-        await ctx.send(f"You can use one of the following channel types:\n{categories}")
-
 
 def setup(bot):
     log.info("[Cog] ServerCommands")
